Fix unit-test for WP 4.7
