+++ tests/php/test-class-customize-snapshot-manager-back-compat.php
@@ -204,6 +204,63 @@ class Test_Customize_Snapshot_Manager_Back_Compat extends \WP_UnitTestCase {
 		$this->assertEquals( 10, has_action( 'customize_save_after', array( $manager, 'publish_snapshot_with_customize_save_after' ) ) );
 		$this->assertEquals( 10, has_action( 'transition_post_status', array( $manager, 'save_settings_with_publish_snapshot' ) ) );
 		$this->assertEquals( 10, has_action( 'wp_ajax_customize_update_snapshot', array( $manager, 'handle_update_snapshot_request' ) ) );
+		$this->assertEquals( 10, has_action( 'customize_preview_init', array( $manager, 'customize_preview_init' ) ) );
+		$this->assertEquals( 10, has_action( 'wp_enqueue_scripts', array( $manager, 'enqueue_frontend_scripts' ) ) );
+	}
+
+	/**
+	 * Test customize preview init.
+	 *
+	 * @see Customize_Snapshot_Manager::customize_preview_init()
+	 */
+	function test_customize_preview_init() {
+		$manager = new Customize_Snapshot_Manager_Back_Compat( $this->plugin );
+		$this->assertFalse( has_action( 'wp_enqueue_scripts', array( $manager, 'enqueue_preview_scripts' ) ) );
+		$manager->customize_preview_init();
+		$this->assertEquals( 10, has_action( 'wp_enqueue_scripts', array( $manager, 'enqueue_preview_scripts' ) ) );
+	}
+
+	/**
+	 * Test enqueue preview scripts.
+	 *
+	 * @see Customize_Snapshot_Manager::enqueue_preview_scripts()
+	 */
+	function test_enqueue_preview_scripts() {
+		$manager = new Customize_Snapshot_Manager_Back_Compat( $this->plugin );
+		$manager->ensure_customize_manager();
+		$manager->init();
+		$handle = 'customize-snapshots-preview';
+		$this->assertFalse( wp_scripts()->query( $handle, 'enqueued' ) );
+		$this->assertFalse( wp_styles()->query( $handle, 'enqueued' ) );
+		$manager->enqueue_preview_scripts();
+		$this->assertTrue( wp_scripts()->query( $handle, 'enqueued' ) );
+		$this->assertTrue( wp_styles()->query( $handle, 'enqueued' ) );
+
+		$after = wp_scripts()->get_data( $handle, 'after' );
+		$this->assertNotEmpty( $after );
+		$this->assertContains( 'CustomizeSnapshotsPreview', join( '', $after ) );
+	}
+
+	/**
+	 * Test enqueue frontend scripts.
+	 *
+	 * @see Customize_Snapshot_Manager::enqueue_frontend_scripts()
+	 */
+	function test_enqueue_frontend_scripts() {
+		$this->plugin->register_scripts( wp_scripts() );
+		$this->plugin->register_styles( wp_styles() );
+		$manager = new Customize_Snapshot_Manager_Back_Compat( $this->plugin );
+		$manager->init();
+		$this->assertFalse( wp_script_is( 'customize-snapshots-frontend', 'enqueued' ) );
+		$manager->enqueue_frontend_scripts();
+		$this->assertFalse( wp_script_is( 'customize-snapshots-frontend', 'enqueued' ) );
+
+		$_REQUEST['customize_snapshot_uuid'] = self::UUID;
+		$manager = new Customize_Snapshot_Manager_Back_Compat( $this->plugin );
+		$manager->init();
+		$this->assertFalse( wp_script_is( 'customize-snapshots-frontend', 'enqueued' ) );
+		$manager->enqueue_frontend_scripts();
+		$this->assertTrue( wp_script_is( 'customize-snapshots-frontend', 'enqueued' ) );
 	}
 
 	/**
